Bug#32555575 SAMPLE SCAN DO NOT PRODUCE ROWS BASED ON ISOLATION LEVEL SET
Issue
=====
As part of histogram sampling, the code was hacked to set the tansaction
isolation level to READ_UNCOMMITTED as the transaction support for
sampling was not required. But the TABLESAMPLE feature being introduced
by the optimizer team uses the same path and requires transactional
support.

Fix
===
Modify the sample interface to add an additional parameter trx_support
which indicates whether transaction support needs to be enabled or not
for the sample read.

RB: 26052
Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>
Reviewed-by: Mithun C Y <mithun.c.y@oracle.com>