Bug#32328576: BOOL PROCESS_BUFFERED_WINDOWING_RECORD: ASSERTION               `!ROW_ADDED && !FOUND_FIRST' FAILED.
Fields of type enum/set are ordered based on numeric comparison.
However, comparison function for range expressions (expression
used for ordering in case of a RANGE frame specification) of a
window function is set based on the result type of the field which is
STRING for enum/set datatype.
As a result, processing of rows for a window frame (to see if a
row is before/after the frame) does not work correctly. A string
comparion might detect a row is before a frame, while the numeric
comparison would have placed the row after. Because of this, we
hit various asserts.

Solution:
Set up integer cache items for enum/set types. Also, setup integer
comparison functions when enum/set types are involved for range
expressions.

Change-Id: I508c7933cee9233e912614fcf01247088748e7b8