Post-push fix for Bug#32221703.
This patch is for mysql-8.0.

Problem
-------
slave_rows_search_algorithms is not set to HASH_SCAN explicitly inside the
test case because of that the jobs running the test with different values failing
due to result content mismatch.

Fix
---
slave_rows_search_algorithms is set to HSAH_SCAN and proper cleanup is done
in the end for all the variables used in the test.

RB#26466