Bug#32929012 remove gcc-4.x workaround for std::regex
Issue
=====

The http server plugin depends on posix regex C-functions to workaround
the missing std::regex support in the libstdc++ of gcc-4.x.

As GCC 4.x isn't supported anymore as C++14 is the base requirement to
build MySQL Router, the corresponding code can be removed and std::regex
be used on all supported compilers.

Change
======

- use std::regex directly
- remove posix_re.h

RB: 26471