Bug#32924198: SIG11 IN HA_INNOBASE::FT_READ|HANDLER/HA_INNODB.CC
Queries which need to sort the results of a full-text index scan could
in some circumstances make the server exit.

When a sort iterator is on top of a full-text search iterator, the
sort iterator turns the underlying full-text index scan into a random
scan once it has sorted all the rows. This is in order to do random
lookups, using row IDs stored in the sort buffer, to fetch values for
required columns that are not in the sort buffer. In doing this, it
clears handler::ft_handler (which is the full-text handler).

If the sort iterator is executed multiple times in a query, it
reinitializes the full-text index scan, which is supposed to turn the
underlying random scan back into a full-text index scan. But it
doesn't restore handler::ft_handler to its original value, and this
causes problems when ft_handler is accessed later.

The bug is fixed by not clearing handler::ft_handler when switching
from full-text index scan mode to random scan mode, which was also
suggested by a TODO in Item_func_match::init_search(). We still need
to tell Item_func_match::val_real() that it must use a different
method of retrieving the match score. It already has a flag for
that (Item_func_match::join_key), so we use that flag instead. And we
also ensure that the flag is restored when the full-text index scan is
reinitialized in FullTextSearchIterator::Init().

The Item_func_match::join_key member variable is renamed to
score_from_index_scan, to make it clearer what it is used for.

Change-Id: I3c4d5e3a716511fc468675b5f142a1788266bb2a