Bug#32932970 GET_UINT_OPTION IS DUPLICATED IN THE CODE.
The function that converts the string from the configuration into
unsigned value is duplicated in the code (routing plugin has its own
version).
Also it can't be used outside of the plugin - like for parsing the
global DEFAULT section options.

This patch moves the generic function to the harness to remove the
duplication and to make this functionality accessible outside of the
plugin code.

RB: 26476