Bug #33119009 IMPLEMENT INNODB LIBRARY FOR LARGE-PAGE ALLOCATIONS WITH 4K-PAGE FALLBACK SUPPORT
This patch introduces library functions that are designed to handle
large (huge) page allocations but which also implement graceful automagic
fallback to regular pages when former ones are not able to be used.
Falling back to regular pages may take place either when the large (huge)
page address space is exhausted (no more huge pages left) or when the
support is explicitly disabled through OS layers or through MySQL specific
flag (--use-large-pages).

This patch builds its grounds on top of the work from:
 Bug #33118362 IMPLEMENT INNODB LIBRARY FOR 4K-PAGE SUPPORT
 Bug #33118309 IMPLEMENT INNODB LIBRARY FOR LARGE-PAGE SUPPORT

Memory instrumentation through PFS, a performance schema storage engine,
is preserved and is an opt-in feature of this library.

Set of new library functions are:
 * ut::malloc_large_page_withkey(fallback_to_normal_page_t)
 * ut::malloc_large_page(fallback_to_normal_page_t)
 * ut::free_large_page(fallback_to_normal_page_t)
 * ut::large_page_allocation_size(fallback_to_normal_page_t)

Reviewed-by: Marcin Babij <marcin.babij@oracle.com>
RB: 26538