BUG#32934479: WHEN "GROUP REPLICATION APPLIER MODULE" IS KILLED, MEMBER MUST CHANGE TO ERROR
Group Replication has a applier module that handles the group
transactions and other messages.
If the thread that runs that module is killed, the member is unable
to handle those messages, that is, unable to run correctly.
It was observed that if that thread was killed, the member state
would continue ONLINE, ignoring the internal errors.

In order to fix the above issue, now when the applier module thread
is killed the member changes to ERROR state, performing all fencing
tasks like leave the group.

RB: 26485