Bug#32721533 CAN'T ALTER TABLE AFTER MOVING THE DATA FILE USING INNODB_DIRECTORIES OPTION
Issue
=====
When a tablespace file is moved offline and the directory in which the file
resides now is made known to the server via the `innodb_directories` option,
ALTER TABLE copy algorithm fails with an unknown storage engine error.

The issue is that there is a rename check as part of the ALTER TABLE statement,
and for that the ibd file is searched in the data directory instead of the
directory it is actually present in, leading to FIL_MISSING error.

Fix
===
Search for the tablespace file in the appropriate directory. This information
is correctly tracked in fil_node_t structure of the corresponding file, so use
the path present in this to search for the file instead of blindly searching
the data directory.

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
RB: 26237