Bug#31636339 MTS IS UNUSABLE; ERRNO 3030, AND WORKER HA_ERR_FOUND_DUPP_KEY OR HA_ERR_LOCK_WAIT_TIMEOUT
This patch is for mysql-5.7.

Background
----------
In the case of binlog_row_image=minimum, we put a minimum number of fields
in the read set and only the modified fields in write sets. As these read
sets and write sets are used in generating the write set hashes and writset
hashes are used to compute the transaction dependencies.

Problem
-------
In this test case, we have a primary key and a secondary key (Unique
key) so it adds only the primary key in the
readset and only the changed field in the writeset such that the
secondary key is not included in the writeset hashes
since it is not there in either of read and write sets.

Due to this transaction dependencies are not calculated correctly on
source and we get an error while executing these transactions parallelly
on replica in a multi-threaded environment.

Solution
---------
If we include secondary key (unique) in the writeset hashes without
including it in the before and after image then transaction dependencies
will be computed correctly and there would not be any failures on replica
due to this.

Fix
---
1. Take temporary backup of readset and writeset and restore these bitmaps after
   setting the bits for unique key (step 2) in order to remove 1-1 correspondence
   between readset and before-image. And this is being done in the
   Binlog_log_row_cleanup class.

2. In case of BINLOG_ROW_IMAGE=MINIMAL, mark all unique key bits in the readset
   and writesets to calculate the transaction dependencies correctly.

3. In the function add_pke, pointer arithmetic code where we are calculating the
   writeset hashes is also modifies same as Mysql-8.0.

RB#25966