Bug#32813547: Segv in sql_optimizer.cc:7415 of MySQL server
The problem manifests in add_key_part() when adding a key lookup
for a reference to a common table expression. This happens because
the key_info array has not been set up for the table. This was supposed
to happen in TABLE_LIST::generate_keys(). This function is supposed to
loop over all TABLE objects referencing the same CTE, however in this
case, one table had been deleted because it belonged to a query
expression object that had been eliminated. The root cause of the
problem is the iterator Derived_refs_iterator that returns nullptr
for a TABLE that is nullptr, thus marking a premature end of the list
of tables.

The fix for the problem is to skip elements of the list where the
TABLE is nullptr.

Notice that patch is also depending on fix for bug#32813554.

Change-Id: Ie53407596733a074cfd6ebaa34b4172964950694