Bug#32065606: Inconsistent results when running the same query
This problem was caused by a combination of a bad nullability
indication for an Item_field object and one change in the fix of
bug#31405780.

The bad nullability might happen for a query that outer-joins
two tables to a third table, and both joins are nested-loop joins with
ref access. The first outer-joined table is a derived table. The second
outer-joined table is a regular table, and it has a join condition with
the first outer-joined table.
Further on, the column in the join condition from the first outer-joined
table is not nullable and the column in the join condition from
the second outer-joined table must be nullable. In this case, the
join condition column from the first outer-joined table is marked as
not nullable but is actually nullable (since it is outer-joined).
The wrong nullability is encountered when we outer-join a derived table.

The code change that enabled the problem was in
TABLE_REF::impossible_null_ref:

-      if ((null_rejecting & 1 << i) && items[i]->is_null()) {
+      if ((null_rejecting & 1 << i) && items[i]->maybe_null &&
+          items[i]->is_null()) {

Although the change by itself is safe, if maybe_null is incorrectly
false, we may have a wrong result. In that case, we will not be
reporting a NULL row, but will rather be attempting to get actual
column values from the outer-joined tables.
Thus, the values reported for this row will depend on the random
contents of the row buffer from previous fetches.

The actual bug seems to be fixed by WL#14333, which did some
updating of column nullability from derived tables.

We strengthen the code by reverting the fix in
TABLE_REF::impossible_null_ref.

A test case is also added for regression testing.

Reviewed by: Chaithra Gopalareddy <chaithra.gopalareddy@oracle.com>