Bug#32924533: Conditions not pushed down to tables inside views or sub queries
When referring columns exposed through a view or a table sub-query,
an Item_ref object is constructed by the optimizer. This Item object
has an 'Item **ref' member which refers the Item(s) the view/sub-query
column is defined over.

It turns out that we lacked handling of Item_ref objects in the
condition pushdown integration, which is added by this patch:

- The Item_ref::traverse_cond() method is implemented.
- Handling of Item::REF_ITEM is added to ndb_serialize_cond()

Note that the 'handling' basically consists of continuing into th e
Item referred by the Item_ref. (Instead of taking the default action
which is setting 'not supported for condition pushdown')

Reviewed by: Martin Skold <Martin.Skold@oracle.com>