Bug#32942327: CAST: ITEM_TYPECAST_UNSIGNED: ASSERTION `!NULL_VALUE || IS_NULLABLE()' FAILED [patch 1/2, comparator]
If an error was raised while evaluating a comparison of two
non-nullable values as strings, the result of the comparison was set
to NULL, even though the metadata of the comparison operator said the
result was non-nullable. The error was correctly returned to the user,
but an assertion would complain about this inconsistency when running
in debug mode.

Fixed by telling Arg_comparator not to set its owner to NULL if the
owner isn't nullable. Arg_comparator already had a flag for this,
called set_null.

Change-Id: If83e70f87b08022de4e81bca40eb6fa14e1feb4c