Bug#32326510 - REPREPARE AT THE BEGINNING OF AN IMPLICIT TRANS WILL CAUSE 1837 ERROR
When autocommit is enabled and gtid_next is set, executing prepared
statement results in an ER_GTID_NEXT_TYPE_UNDEFINED_GROUP error.

In autocommit mode if gtid_next is set then during statement
transaction commit or rollback gtid_next is invalidated. Even for a
prepared statement execution gtid_next is invalidated during transaction
commit or rollback. But for a prepared statement, if re-prepare is
triggered (due to dependent table metadata change) then prepared
statement execution returns with re-prepare error and transaction
is rolled back. During transaction rollback gtid_next is invalidated.
On execution of re-prepared statement, an error is reported for
incorrect gtid_next value.

This issue exists with stored programs too. LEX instance of a
SQL statement in stored program is cached. If metadata of a table
used by SQL statement changes then re-prepare is triggered for
stored program statements and gtid_next is invalidated.

To fix this issue, gtid_next update during transaction rollback is
skipped for the prepared statement and stored program statements
if re-prepare is triggered. Same gtid_next value is reused for
executing re-prepared statement with this change.

Change-Id: I5eda7a056f0b2083af316b92f7ecc2d9185c2576