Bug#32877703: REWRITE JSON_LENGTH(DOC, PATH) TO JSON_LENGTH(JSON_EXTRACT(DOC, PATH))
Post-push fix for doxygen warnings.

item_create.cc:2196: warning: end of file while inside a group

The end tag of a defgroup was inside an anonymous namespace, and
wasn't seen by doxygen. Fixed by expanding the anonymous namespace to
also cover the defgroup.

item_create.h:199: warning: argument 'a' of command @param is not
found in the argument list of create_func_cast(THD *thd, const POS
&pos, Item *arg, const Cast_type &type, bool as_array)

The declaration and the definition used different names for one of the
parameters. Fixed by using the same name in both.

Change-Id: Iae59c67df7f61a640a96ffd71830ba7d8c8f58f4