Bug#32918240: Server failure, subqueries with specific conditions Bug#32787861: Table expression in prepared statement fails
The contextualization rules for query expression were missing a rule
to contract two query specifications where the inner one has an
ORDER BY clause and the outer one has a LIMIT clause. Due to this,
the query expression optimization attempts to access a TABLE object
that was created during preparation but has not been opened.

Fixed by adding the rule in contextualization.

Reviewed by: Gleb Shchepa <gleb.shchepa@oracle.com>

Change-Id: Id2ff5cd4a33ec81f43114c5e4bdf40e6269ea065