Bug#32891703: UDF ERROR MESSAGE TRUNCATED
When UDF initializing failed, the result error message
could contain incomplete (that is to say, truncated)
diagnostics.

ER_CANT_INITIALIZE_UDF is defined as

  eng "Can't initialize function '%-.192s'; %-.80s"

which of course limits the enclosed message to 80 characters.
It turns out that this actually affects rather a lot of
test cases (or rather, test .results), which we endeavor to
fix with this patch, alongside the actual message template
which we make a little more generous in its dimensions.

Change-Id: I2c95e565792c9da2d4ab77029464e43e39173660