Bug#32803211: 8.0.22+ triggers fail in Item::mark_field_in_map
The problem was that when INSERT ... SET ... statement or UPDATE
statement was used in a trigger we have accepted reference to column
in the table on which trigger was defined (using NEW or OLD designator)
as an assigment target in its SET clause. This always led to
assertion failures in debug builds and sometimes server failures
in production builds.

The problem stemmed from two facts: 1) INSERT and UPDATE statement code
expects only references to their target table columns as assignment
targets of the SET clause 2) we didn't block usage of trigger
fields as target for such assignment, due to omission which has
happened during early days of implementation of triggers.

This patch solves the problem by rejecting trigger fields as
SET clause assignment targets in both INSERT and UPDATE.

This is an incompatible change which breaks rare user's triggers
which use this syntax (which might have been working somehow due
to luck). Such triggers produce an error during attempt to execute
them.

However, the new behavior is in line with both our documentation
and SQL standard, and supporting old non-standard and non-documented
behavor will require quite some efforts.
The fix for the problem is to reject trigger fields as SET clause
assignment targets in INSERT and UPDATE statements.

Reviewed by: Dmitry Lenev <Dmitry.Lenev@oracle.com>