Bug #32919524 	CONTRIBUTION: BACKPORT BUG#28949700: SUBOPTIMAL USE OF STRING::RESERVE() IN ...
Contributed by Annirudh Prasad

Backport of BUG#28949700 SUBOPTIMAL USE OF STRING::RESERVE() IN
WRAPPER_TO_STRING().

The one-argument version of String::reserve() increases the capacity
of the String object linearly and causes a high number of memory
allocations when converting certain JSON documents to text
representation.

This patch changes wrapper_to_string() so that it uses the
two-argument version of String::reserve() instead, which enables
exponential increase of the capacity and reduces the number of memory
allocations.

The patch includes some additional small improvements to make the
conversion from JSON to text go faster:

- When appending a string literal, provide the length as an argument
  to avoid calls to strlen().

- When appending two-character string literals, instead append the two
  characters individually (append(char) is inlined, append(const
  char *, size_t) is not).

Change-Id: I5ee36791d07c6525918b2ae0e5919ce47f05404e