Bug #33118362 IMPLEMENT INNODB LIBRARY FOR 4K-PAGE SUPPORT
This patch introduces library functions that are designed to handle
page-aligned allocations. Normally, most CPUs support and default
to 4K pages as the smallest memory-addressable unit but some architectures
allow for this unit to be configured to be larger. So, for example aarch64
supports normal page sizes to be 4K, 16K or 64K. Similar applies for some
other architectures such as ppc64, sparc64 or mips64.

In contrast to the functions implementing page-aligned allocations
provided by Unix and/or Windows, these ones have built-in support for
memory instrumentation that is implemented on top of PFS, a performance
schema storage engine. Memory instrumentation is an opt-in feature.

Set of new library functions are:
 * ut::malloc_page_withkey
 * ut::malloc_page 
 * ut::free_page
 * ut::page_allocation_size

This set of library functions will allow for more elegant implementation
of the follow-up library whose goal is to allocate memory through
huge-page support but which is able to gracefully fallback to 4K page-aligned
allocations in the event of being unable to back the allocation up from
huge-pages (e.g. huge-page exhaustion).

Reviewed-by: Marcin Babij <marcin.babij@oracle.com>
RB: 26537