Bug#32960158: Resolver does not stop after first error
Generally, the resolver would abort the analysis and bail out after
seeing an error in a statement. This is a simple principle and lets
us carry a minimum of context information during analysis.
However, in the case of duplicate column analysis, the resolver
continues until the end of the column list, possibly adding multiple
error messages to the diagnostics object. However, only one of these
messages are returned as the outcome of the analysis, the SHOW WARNINGS
command is required to list all errors.

Fixed by skipping analysis after first error is encountered.

Reviewed by: Knut Anders Hatlen <knut.hatlen@oracle.com>

Change-Id: Ieabfc12d2a26f0f9d84cd31de02840f1dec313bf