WL#14388 Push condition referring fields from an SPJ ancestor table
Check for catching that entire predicate term could be const-folded to a boolean true/false value
was not entirely correct. Code for this added by WL#12956 only catched the case where a 'bool_func'
(A comparison condition using =, <>, <, >, >=, <=) was const-folded.

This patch replace the original check for this with a more general check, catching all
const-folded expression where a 'COND_ITEM' was expected.

Also re-record a few result files showing the effect of this change.

Reviewed by: Arnab Ray <arnab.r.ray@oracle.com>