WL#14388 Push condition referring fields from an SPJ ancestor table
Refactoring of class Ndb_item in ha_ndbcluster_cond.cc, no functional change.

This WL will later need to introduct a 'Param' variant of the Ndb_item.
In preparation for that we take the oportunity to first clean up
the implementation and usage of the class Ndb_item.

Corrently the Ndb_Item had different C'tor variants for creating
either a VALUE, FIELD. FUNCTION, or END variant of the Ndb_Item.

This patch change the Ndb_item to a pure base class and introduce
derived classes Ndb_func, Ndb_field, Ndb_value and Ndb_end_cond
for the VALUE, FIELD. FUNCTION, or END variants.

Specific methods used exclusively for the different variants are
moved into the derived classes.

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>