WL#14388 Push condition referring fields from an SPJ ancestor table
Fix an issue in setting up the mandatory 'm_ancestor' dependencies
between (outer-) join nests. Actually a bug in WL#7636
'Improve OUTER JOIN capabilities of SPJ'. As we have been unable
to find any failing queries due to this bug, except an incorrect
ancestor (join-child) in the query plan changed by this patch,
we didn't file a bug for this. However, upcomming patches
related to this WL would have failed if not fixed now.

Background: There is a requirement is the SPJ code that
all ancestor references from a SPJ child should either refer
a parent table within its own inner-join-nest, or if tables from
an upper-nest is referred, all such upper-nest referrences from
a join-nest has to refer the same upper-table.

In order to implement checking and enforcing of this rule, we
will add all references to tables outside of this nest to the
'm_tables[table.m_first_inner].m_ancestors' table map. This
m_ancestors set is the used to check for conflicting sets
of upper-ancestor references, and later used when we choose
the actuall parent table to refer.

This patch fix an issue where we may set up the m_ancestors set
of tables incorrectly.

In the changed MTR result we find that the 'table3 left join table4'
did not refer to the same parent table from its upper-nests.
(table1 vs tables2). With this patch table2 becomes the common parent.

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>