Bug#32979815: WL#14422: ASSERTION `!(USED_TABS & (~READ_TABLES & ~FILTER_FOR_TABLE))' FAILED
When the hypergraph optimizer chooses to materialize an expression, it
rewrites the item tree by replacing the materialized expression with a
reference to a column in a temporary table. When doing so, it does not
update the item's used_tables(), so that parents of the replaced item
will still report that they access the original base tables instead of
the temporary table. This inconsistency can lead to assert failures.

So far, problems have been seen only with queries that use full-text
search. These queries are not yet supported by the hypergraph
optimizer, but we fix it now to avoid problems when support for
full-text search is enabled.

The bug is fixed by calling Item::update_used_tables() from
ReplaceMaterializedItems() after the item tree has been modified.

Change-Id: I55faa8ac2eed4c93b8f9b1a01885c3b541a0ead0