WL#14388 Push condition referring fields from an SPJ ancestor table
Enable SPJ block to create parameters for pushed conditions, and append
these to the attrInfo section.

Use the (upto now unused) QueryTree option NI_ATTR_LINKED. When set the
internal SPJ query context will be set up to 'construct' the attrInfo parameter
based on the contents of incomming ancestor rows. (Thus the attrInfo
is 'linked' to these). Note that we also has 'linked' key values
when SPJ create bound-keys for accessing index tables.

SPJ partly already had code for handling T_ATTRINFO_CONSTRUCTED, it
has never been used though. It turned out that the implementation of
it was entirely missing for scanFrag_parent_row(), while it was
implemented in lookup_parent_row()

Note that for (index-)scans we will need a parameter to be generated for
each range we will scan (or, for each parent row). This is similar to
index scans also requiring a range-key to be produced for each parent
row. In order to navigate from one parameter to the next, we extended
the existing (and unused) parameter format in the attrInfo to preceed
each parameter with a 'parameter length' word - See patch for
change made to T_ATTRINFO_CONSTRUCTED in lookup_parent_row()

Parameters are constructed into the m_paramPtrI memory section
when parent rows arrives and calls into *_parent_row() method.
When all parent rows for a batch has arrived we will have the full
set of parameters, and 'send' will be called. 'send' will then
append the parameter section to the attrInfoPtrI-section and update
the section[4] header accordingly. (Where TUP will locate the
parameter as part of executing the interpreter code)

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>