WL#14388 Push condition referring fields from an SPJ ancestor table
Enhance TUP and LQH to handle attrInfo built with multiple parameters. (index scans)

When LQH execute a request requiring TUP to execute interpreter code, it
need to tell TUP to copy out the interpreter code into linear memory (cinbuffer[])
from the supplied attrInfo before the interpreter code can be executed.
For this it will call 'c_tup->copyAttrinfo'.

TUP code was already prepared for handling interpreter code refering a
parameter stored in the attrInfo as well (Never used though). That
parameter was then simply copied into linear memory by copyAttrinfo,
together with the interpreter code.

Introducing multiple parameters, as required when multiple 'bounds' are
specified in a single scan request, needed this logic to be enhanced.
(Note that the previous SPJ patch contains changes to the attrInfo
parameter format; a paramLen is now preceeding each parameter in the
attrInfo)

This patch change the copyAttrinfo such that first copy out
the attrInfo *excluding* the parameters. Then it navigate to the
correct paramNo to be used for the current 'bound' and copy
out only *that* parameter. (For a lookup request the paramNo will
be used, such that the first, and only, parameter is specified)

The new method Dbtup::nextAttrInfoParam() is introduced which is
used by LQH to inform TUP that it should advance to the
next paramNo.

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>