WL#14388 Push condition referring fields from an SPJ ancestor table
Introduce required changes to the pushed condition handler interface
to allow pushed conditions to refer Field from an ancestor table
in the same pushed query.

Note that we already allows Fields from previous tables to be referred
if they are *not* part of a pushed join. In these cases a Field from
a previous table will be 'const at execute', thus we can already when
the interpreter code is generated fetch the Field value and handle it as
a const value. Note that the bitmap 'const_expr_tables' contain the tables
which may be handled in this way.

For a pushed query all the interpreter code is generated prior to the first table
in the pushed query is sent for execution, even for child tables far down
in the query plan. Thus no column values has been retrieved for the Fields being
referred yet. Instead we will let SPJ construct a parameter set to send along with
the interpreter code when the ancestor values are being returned to SPJ.

This patch introduce the Ndb_param class as a subclass of Ndb_Item.
(Already has Ndb_field, Ndb_value and Ndb_func as other variants)
The Ndb_param will describe a Field which is used to construct a
parameter, which is then referred by any of the other Ndb_item constructions.

Note that we intend to allow only one of the two operands to be a 'parameter'.
Furthermore, if there is a parameter, the other operand has to be a Field
reference to 'this' table. (A natural limitation as the condition should
have been pushed further down in the query if this does not hold.)

In order to check and enforce the above limitation the
'expect*_field_from_tables()' lookahead methods are enhanced with
a bitmap of tables which we allowed to refer a Field from, either as a
Ndb_field or Ndb_param. This is used such that when we have found an
Field needed to be handled as a parameter, we only expect the next
operand to be a Field from 'this_table'

Generate of interpreter code has been enhances to handle the Ndb_param,
where it will use the previously introduced NdbScanFilter::cmp_param()
method.

A special situation had to be handled in cases where a condition
were prepared (serialized) for pushing with a parameter (Thus part
of a pushed join,) and we later encountered a situation disallowing
the query push. As only SPJ execution support/need parameter generation,
we need code generation to handle the Ndb_param as a Ndb_value.
(The Field value *can* be retrieved as a const-value in these cases)
The argument 'param_is_const' was introduced to handle these (rare)cases

Note that we still assert on that 'param_expr_tables == 0'
(Even added one more assert). Thus we will still not allow
the Ndb_param-functionality in this patch be be used. (there are some
missing pieces left out for next patch)

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>