Bug #32991055: HYPERGRAPH OPTIMIZER: CREATE FUNCTIONAL DEPENDENCIES FROM AGGREGATES [pre-fix, noclose]
Fix an issue where we would not always reach new group NFSM states,
since they would be pruned away in CouldBecomeInterestingOrderings().

An example issue goes as follows (also unit tested):

Assume we have FDs {b} → a and a = c. Now, can we create the grouping
{b,c} from the ordering (b)? Under our rules, it is never interesting
to extend an order to create a grouping, so we'd first need to convert
to {b}. Then, we'd have to apply {b} → a to get to {b,a}. However,
since we always canonicalize groupings by sorting them, we immediately
end up with {a,b} instead. And since we compare groupings as orderings
in CouldBecomeInterestingOrderings(), we don't understand that {a,b}
could become {b,c} through the a=c equivalence, since we only compare
a with b.

We fix this by changing the canonical sorting of a group; we put
equivalent items together, so that the {b,c} grouping is instead
represented as {c,b}.

This is much more visible when we have groupings to extend, e.g.,
we first GROUP BY, then apply FDs on the result, and then use DISTINCT,
which is why it shows up in the next patch (which adds FDs for
aggregates).

Change-Id: I2a422cc8dd5cf19794af870fa1b4c416baa64f44