Bug#32855634: Assert(m_deque == other.m_deque) in mem_root_deque.h
This problem may occur when we have a stored procedure containing
an IF statement with an EXISTS function inside, which e.g. performs
a JOIN of multiple tables. If that logical function needs a reprepare,
e.g. because it involves a table that was deleted and recreated between
two executions, we may see a failed optimization.

The reason for this is that in order for the logical function to be
reprepared, the text string is first prefixed with 'SELECT '. Thus,
the parser will parse this as a SELECT statement. However, the resolver
will see a query block structure that is unexpected, and not call
the function Query_block::apply_local_transforms() which (among other
things) normalizes the structures with tables to be joined, hence the
later failure. For a logical function, the resolver expects to see
the SQL command SQLCOM_END (rather than SQLCOM_SELECT).

There is actually a feature in stored procedure handling to fix
such problems. A class function adjust_sql_command() may be used to
adjust the SQL command type after parsing. The fix to this problem is
thus to implement the function adjust_sql_command() for the class
sp_lex_branch_instr (which implements the IF statement), and adjust
the SQL command type to SQLCOM_END.

The problem exists for other procedural statements, such as CASE,
WHILE and REPEAT. Test cases have been added for all these.

Reviewed by: Gleb Shchepa <gleb.shchepa@oracle.com>

Change-Id: Ib72ae28a32207891c522b44afdee30e2a1a6e5db