Bug#32890918: Writing the Windows EventLog would occasionally fail
When converting the message to be logged to UCS-2 for Windows, the
conversion routine could read beyond the end of the input string.
Patch prevents this, adds comments to better explain the idiosyncrasies
of our Windows logging, and renames variables to hopefully prevent
uncertainty in the future.

This patch comes with a big thank you to Shane Bester for dumps
and other helpful files when the issue was elusive on my machine,
and to Daniel Blanchard who went above and beyond creating a much
optimized proof once I had a patch.

Change-Id: Id28d09dcc5a51caf67d966034d485d28eed5c84e