Bug#32816305: Prepare type checking error
After WL#9384, there was a regression in type resolving for a NULLIF
function with a dynamic parameter. When the first argument to
the function is a dynamic parameter, the type was determined based on
the context, e.g if NULLIF was used in an assignment to a DATE column,
it would be assigned a DATE type. However, this is not compatible with
how type resolving used to work. It is better to take the type of a
dynamic parameter from the other argument, like a comparison operation
does, and only deduce the type from the context when both arguments
to NULLIF are dynamic parameters. This ensures there is no type
conversion involved in the comparison.

The new type determination procedure is thus:
- If both arguments are dynamic parameters, their type will be
  determined from the context. E.g, if NULLIF is used in an assignment
  to a CHAR column, both parameters are assumed to be of type CHAR.
- If one of the arguments is a dynamic parameter, it will assume
  the type of the other argument.

A tiny cleanup: agg_arg_charsets_for_comparison() used to be
called always, now it is only called when converting result from a
temporal type to a string type, since Item_bool_func2::resolve_type()
would call this function.

Change-Id: Ifc9899fd0774c406ea89ababb6945b3e0591ee77