Bug #32970679: REPLACE MYSQL::NULLABLE WITH STD::OPTIONAL
Post-push fix: Fix gis.ddl. The error message sent an uint32_t to a
function with the printf format %lu, which is not the same, and yields
undefined behavior. This worked fine for most cases, but after we
switched Mysql::Nullable to std::optional, Clang would optimize this
in a different way that would add 2^32 to the printed SRID
(ie., it would not clear the upper 32 bits, since the ABI said it does
not have to).

Change-Id: I14c71588780064439d893060368b9b626e6bf8d6