Bug #33003581: HYPERGRAPH: ASSERTION `FALSE' FAILED|SQL/ITEM_CMPFUNC.CC
When we replace a subtree with a ZERO_ROWS access path due to impossible
join conditions, we also need to remove equijoin conditions referring to
that subtree, or the hash join iterator will be sad.

This is essentially a forward-port of bug #31563414 (which fixed the
same issue in the old access path construction) to the hypergraph
join optimizer.

Change-Id: I568056be19450b530617efacd6641b305f7d2df2