Bug #32644631: PRELIMINARY FIXES FOR WL #14419 [force_stable_sort, noclose]
Remove the redundant filesort parameter force_stable_sort. It was
intended to guarantee a consistent total ordering of rows, but stable
sorting does not actually do that, and filesort isn't guaranteed stable
even if you sort each partition stably (spill-to-disk has no such
guarantees).

Furthermore, we already had other code
(Window::reorder_and_eliminate_sorts()) which did a much better job at
tackling this problem -- for the hypergraph optimizer, the interesting
order code will have the same effect. It is not a complete solution,
though, so add a TODO to note so.

Change-Id: I74dff8be7ceedbcec8d46cc35576b4138a3f2ec3