Bug #32808809 	INNODB: DEADLOCK BETWEEN FIL_SHARD::OPEN_FILE, FIL_OP_REPLAY_RENAME ON WINDOWS
Keeping file opened outside fil_system without SHARE_WRITE and trying to acquire the fil_shard mutex may lead to a deadlock with a thread that wants to access this file, when it acquired the fil shard mutex and wants to open the file.

`Datafile::validate_first_page()` is now closing the file before it would acquire the fil shard mutex. It is not being used afterwards, but flags and other info that was acquired before closing and which is still present even after the file is closed.

RB#26506