WL#13594 Deprecate character set ucs2 and collations
 If character set ucs2 is specified in a SQL statement, issue a warning:

 'ucs2' is deprecated and will be removed in a future release. Please use
 utf8mb4 instead

If collation is specified for character set ucs2 in a SQL statement,
issue a warning, e.g.

 'ucs2_persian_ci' is a collation of the deprecated character set ucs2.
 Please consider using utf8mb4 with an appropriate collation instead.

If character set ucs2 is specified when starting the server,
issue a warning in the log, e.g.

 --character-set-server: The character set ucs2 is deprecated and will be
 removed in a future release. Please consider using utf8mb4 instead.

If a collation is specified for character set ucs2 when starting the
server (FIXME: other use cases?), issue a warning in the log, e.g.

 --collation-server: 'ucs2_persian_ci' is a collation of the deprecated
 character set ucs2. Please consider using utf8mb4 with an appropriate
 collation instead.

The WL introduces three new warnings:
In messages_to_clients.txt:

   ER_WARN_DEPRECATED_UCS2_COLLATION

In messages_to_error_log.txt:
   ER_WARN_DEPRECATED_UCS2_CHARSET_OPTION
   ER_WARN_DEPRECATED_UCS2_COLLATION_OPTION

Change-Id: I1586565cd82051b6106d67a6b5d84cbe52ddbd14